requirements.txt
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.1
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
regex==2022.7.25
six==1.16.0
SQLAlchemy==1.4.40
textblob==0.17.1
tqdm==4.64.0
Werkzeug==2.2.2
zipp==3.8.1
